Client/Sever Message Listener.
Added 2 methods for Carla to connect to program based on Carla's output. Either using the program as a server or a client and sending or listening to Carla. Then Artery was set up to act as a server as well. Additionally redundant files were removed. Carla API bridge was setup as an alternative for manipulating Carla files if needed, and a converter was added to better interpret potential JSON files from Artery.